requirements.txt
pandas
xgboost
scikit-learn
requests
MLB-StatsAPI
joblib